+++ inc/ingame/ingame_structure_controller.h
@@ -3,6 +3,7 @@
 #include "common/vector2d.h"
 
 #include "entity/beacon.h"
+#include "entity/mana_shard.h"
 #include "entity/monster_nest.h"
 
 #include "interface/window.h"
@@ -11,12 +12,16 @@
 #include <vector>
 
 class IngameLevelDefinition;
+class IngameManaPool;
 class IngameMap;
 
 class IngameStructureController
 {
 private:
+    IngameManaPool& manaPool;
+
     std::vector<MonsterNest> monsterNests;
+    std::list<ManaShard> manaShards;
     std::list<Beacon> beacons;
     vector2d<int> tileStatic;
 
@@ -26,12 +31,13 @@ private:
     void destroyBeacon(Beacon* pBeacon);
 
 public:
-    IngameStructureController(const IngameLevelDefinition& level);
+    IngameStructureController(IngameManaPool& manaPool_, const IngameLevelDefinition& level);
 
     void render(struct _fbg* pFbg, const Window& window) const;
 
     void tickStructures(IngameMap& map, int frames);
 
+    ManaShard& addManaShard(int x, int y, int size, double mana, double shell, bool corrupted);
     Beacon& addBeacon(int x, int y);
 
     bool checkStaticBeacons(int x, int y, int width, int height);
@@ -39,5 +45,5 @@ public:
     std::vector<MonsterNest>& getMonsterNests() { return monsterNests; }
 
     std::vector<Targetable*>& getTargetableStructuresWithinRangeSq(std::vector<Targetable*>& targets,
-        float y, float x, float rangeSq);
+        float y, float x, float rangeSq, bool isStructureTarget);
 };